Remove theme, create DarkModeToggle component
