--- components/Theme/dark.js
@@ -1,19 +1,0 @@
-export default {
-  text: {
-    primary: 'yellow',
-    secondary: '#e3e4e8',
-    tertiary: '#a9abb6',
-    quarternary: '#6c6f7e',
-    placeholder: 'rgba(145,148,161,0.5)',
-    onPrimary: '#050505',
-  },
-  accent: {
-    blue: '#0A84FF',
-    green: '#32D74B',
-    orange: '#FF9F0A',
-    red: '#FF453A',
-    yellow: '#FFD60A',
-    purple: '#BF5AF2',
-    pink: '#FF375F',
-  },
-};
