--- components/Theme/light.js
@@ -1,19 +1,0 @@
-export default {
-  text: {
-    primary: '#fff',
-    secondary: '#2f3037',
-    tertiary: '#525560',
-    quarternary: '#9194a1',
-    placeholder: 'rgba(82,85,96,0.5)',
-    onPrimary: '#ffffff',
-  },
-  accent: {
-    blue: '#007AFF',
-    green: '#34C759',
-    orange: '#FF9500',
-    red: '#FF3B30',
-    yellow: '#FFCC00',
-    purple: '#AF52DE',
-    pink: '#FF2D55',
-  },
-};
