+++ pages/index.jsx
@@ -1,36 +1,19 @@
-/* eslint-disable react/button-has-type */
 /* eslint-disable react/prop-types */
 import * as React from 'react';
 import Head from 'next/head';
-import useDarkMode from 'use-dark-mode';
 import Header from '../components/header';
 import List from '../components/list';
+import DarkModeToggle from '../components/dark-mode-toggle';
 
 import { categories, mapping } from '../data/posts-manifest';
 
 import '../styles/main.css';
 
 function Home() {
-  const darkMode = useDarkMode(false);
   return (
     <>
-      <button onClick={darkMode.toggle}>
-        {darkMode.value
-          ? (
-            <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="feather feather-sun">
-              <circle cx="12" cy="12" r="5" />
-              <line x1="12" y1="1" x2="12" y2="3" />
-              <line x1="12" y1="21" x2="12" y2="23" />
-              <line x1="4.22" y1="4.22" x2="5.64" y2="5.64" />
-              <line x1="18.36" y1="18.36" x2="19.78" y2="19.78" />
-              <line x1="1" y1="12" x2="3" y2="12" />
-              <line x1="21" y1="12" x2="23" y2="12" />
-              <line x1="4.22" y1="19.78" x2="5.64" y2="18.36" />
-              <line x1="18.36" y1="5.64" x2="19.78" y2="4.22" />
-            </svg>
-          )
-          : <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="feather feather-moon"><path d="M21 12.79A9 9 0 1 1 11.21 3 7 7 0 0 0 21 12.79z" /></svg>}
-      </button>
+      <DarkModeToggle />
+
       <Head>
         <title>Cosmonauts World</title>
       </Head>
